route53 role depends on kiam role

+++ controllers/awsmachinetemplate_controller.go
@@ -111,7 +111,8 @@ func (r *AWSMachineTemplateReconciler) Reconcile(req ctrl.Request) (ctrl.Result,
 				return ctrl.Result{}, err
 			}
 		}
-		if r.EnableRoute53Role {
+		// route53 role depends on KIAM role, so it will be crated only if both roles are enabled
+		if r.EnableKiamRole && r.EnableRoute53Role {
 			err = iamService.DeleteRoute53Role()
 			if err != nil {
 				return ctrl.Result{}, err
@@ -135,7 +136,8 @@ func (r *AWSMachineTemplateReconciler) Reconcile(req ctrl.Request) (ctrl.Result,
 				return ctrl.Result{}, err
 			}
 		}
-		if r.EnableRoute53Role {
+		// route53 role depends on KIAM role
+		if r.EnableKiamRole && r.EnableRoute53Role {
 			err = iamService.ReconcileRoute53Role()
 			if err != nil {
 				return ctrl.Result{}, err
